Move exitcommand to its own class.
