+++ org.metaborg.spoofax.shell.core/src/test/java/org/metaborg/spoofax/shell/client/ReplTest.java
@@ -17,6 +17,7 @@ import java.io.OutputStream;
 import org.junit.Test;
 import org.metaborg.core.context.IContext;
 import org.metaborg.spoofax.shell.commands.CommandNotFoundException;
+import org.metaborg.spoofax.shell.commands.ExitCommand;
 import org.metaborg.spoofax.shell.commands.ICommandInvoker;
 
 import com.google.inject.Guice;
@@ -124,7 +125,7 @@ public class ReplTest {
             setUpExit();
 
             // Stub the invoker so that it returns an exit command which we can spy on.
-            Repl.ExitCommand exitCommandMock = spy(new Repl.ExitCommand(() -> repl));
+            ExitCommand exitCommandMock = spy(new ExitCommand(() -> repl));
             when(invokerMock.commandFromName("exit")).thenReturn(exitCommandMock);
 
             repl.run();
